pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Cloudey.app backend - FastAPI + LangGraph + OCI SDK"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "langgraph>=0.0.1",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "oci>=2.112.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "cryptography>=41.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "requests>=2.31.0",
    "boto3>=1.28.0",
    "redis>=5.0.0",
    "apscheduler>=3.10.0",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary,pool]>=3.2.0",
]
